cleaning up modules
